backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
supabase==2.0.2
motor==3.3.2
pymongo==4.6.0
sqlalchemy==2.0.23
asyncpg==0.29.0

# AI and Machine Learning
openai==1.3.7
google-generativeai==0.3.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
langchain==0.0.350
langchain-openai==0.0.2

# Financial Analysis
yfinance==0.2.28
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2
statsmodels==0.14.0
pandas-ta==0.3.14b0
ta==0.10.2
finta==1.3

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.1.0
pytesseract==0.3.10
python-magic==0.4.27

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# HTTP and API
httpx==0.25.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
